Job functionality

+++ code/player/Player.Money.cs
@@ -5,14 +5,14 @@ namespace Cinema;
 public partial class Player
 {
     // The players current money
-    [Net] public int Money { get; set; }
+    [Net] public double Money { get; set; }
 
     /// <summary>
     /// Checks if a player can afford a certain amount of money.
     /// </summary>
     /// <param name="amount">How much money</param>
     /// <returns>True if money >= the amount</returns>
-    public bool CanAfford(int amount)
+    public bool CanAfford(double amount)
     {
         return Money >= amount;
     }
@@ -21,7 +21,7 @@ public partial class Player
     /// Adds money to the player's money.
     /// </summary>
     /// <param name="amount">The amount to add</param>
-    public void AddMoney(int amount)
+    public void AddMoney(double amount)
     {
         Money += amount;
     }
@@ -30,7 +30,7 @@ public partial class Player
     /// Takes money from the player.
     /// </summary>
     /// <param name="amount">How much to take</param>
-    public void TakeMoney(int amount)
+    public void TakeMoney(double amount)
     {
         Money -= amount;
 
@@ -42,7 +42,7 @@ public partial class Player
     /// Sets the amount of money the player has.
     /// </summary>
     /// <param name="amount">How much money</param>
-    public void SetMoney(int amount)
+    public void SetMoney(double amount)
     {
         Money = amount;
     }
